Hello world service
Without this patch there is no end-to-end service to exercise the web
application framework provided by trapperkeeper.  This patch addresses
the problem by making a trivial web application service operate using
the following kickoff command:

    lein trampoline run --config resources/config.conf

The bootstrap configuration doesn't need to be specified on the command
line because it should stay the same across dev/test/prod and the
resources/ folder is already on the classpath.  This patch moves
bootstrap.cfg from dev-resources to resources because dev-resources is
prior to resources on the class path.