Use webserver $PORT when defined in the environment
Without this patch the application fails to run in Heroku because the
application binds to the wrong TCP port.  Statically defining the TCP
port in a configuration file is the root cause of this problem because
the TCP port is dynamically allocated by Heroku on deployment of the
application.  The port is passed to the application via the PORT
environment variable.

This patch addresses the problem by using the
[override-webserver-settings!][override] API method to pass the value
from the environment, if defined, to the WebserverService.

[override]: https://github.com/puppetlabs/trapperkeeper-webserver-jetty9#override-webserver-settings